Update Weizmann training code to handle learning of foreground masks.

+++ datasets/weizmann.py
@@ -131,8 +131,8 @@ def preprocess_mask(mask):
     side = min(height, width)
     x0 = (width - side)//2
     y0 = (height - side)//2
-    # Crop to central square
-    mask = np.array(mask[y0:y0+side, x0:x0+side, :])
+    # Crop to central square, convert to float
+    mask = np.array(mask[y0:y0+side, x0:x0+side, :], dtype=np.float64)
     # Transpose to (time, rows, cols)
     mask = np.transpose(mask, (2,0,1))
     # Resize to 64 by 64
